pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ts-haystack"
version = "0.1.0"
description = "TS-Haystack: A semi-synthetic benchmark for retrieval and reasoning over long time series"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AI-X-Labs" },
]
keywords = ["time-series", "benchmark", "retrieval", "reasoning", "haystack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy",
    "polars",
    "pyarrow",
    "pandas",
    "pyyaml",
    "tqdm",
    "joblib",
    "scipy",
]

[project.optional-dependencies]
download = [
    "huggingface-hub",
]
cot = [
    "openai",
    "matplotlib",
    "seaborn",
    "Pillow",
]
dev = [
    "pytest",
    "ruff",
]
all = [
    "ts-haystack[download,cot,dev]",
]

[project.urls]
Homepage = "https://github.com/AI-X-Labs/ts-haystack"
Repository = "https://github.com/AI-X-Labs/ts-haystack"

[tool.hatch.build.targets.wheel]
packages = ["src/ts_haystack"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"
